refactor(design): rework button example import behavior
