--- libs/design/button/examples/src/examples.ts
@@ -1,15 +1,0 @@
-import { BasicButtonComponent } from './basic-button/basic-button.component';
-import { IconButtonComponent } from './icon-button/icon-button.component';
-import { RaisedButtonComponent } from './raised-button/raised-button.component';
-import { SizeableButtonComponent } from './sizeable-button/sizeable-button.component';
-import { StrokedButtonComponent } from './stroked-button/stroked-button.component';
-import { UnderlineButtonComponent } from './underline-button/underline-button.component';
-
-export const BUTTON_EXAMPLES = [
-  BasicButtonComponent,
-  IconButtonComponent,
-  RaisedButtonComponent,
-  SizeableButtonComponent,
-  StrokedButtonComponent,
-  UnderlineButtonComponent,
-];
