+++ libs/design/button/examples/src/icon-button/icon-button.component.ts
@@ -5,6 +5,12 @@ import { faPlus } from '@fortawesome/free-solid-svg-icons';
   // eslint-disable-next-line @angular-eslint/component-selector
   selector: 'icon-button',
   templateUrl: './icon-button.component.html',
+  styles: [`
+    button[daff-button],
+    a[daff-button] {
+      margin: 4px;
+    }
+  `],
 })
 export class IconButtonComponent {
   faPlus = faPlus;
