+++ libs/design/button/examples/src/public_api.ts
@@ -1,10 +1,21 @@
-export { BasicButtonComponent } from './basic-button/basic-button.component';
-export { IconButtonComponent } from './icon-button/icon-button.component';
-export { RaisedButtonComponent } from './raised-button/raised-button.component';
-export { SizeableButtonComponent } from './sizeable-button/sizeable-button.component';
-export { StrokedButtonComponent } from './stroked-button/stroked-button.component';
-export { UnderlineButtonComponent } from './underline-button/underline-button.component';
+import { BasicButtonComponent } from './basic-button/basic-button.component';
+export { BasicButtonModule } from './basic-button/basic-button.module';
+import { IconButtonComponent } from './icon-button/icon-button.component';
+export { IconButtonModule } from './icon-button/icon-button.module';
+import { RaisedButtonComponent } from './raised-button/raised-button.component';
+export { RaisedButtonModule } from './raised-button/raised-button.module';
+import { SizeableButtonComponent } from './sizeable-button/sizeable-button.component';
+export { SizeableButtonModule } from './sizeable-button/sizeable-button.module';
+import { StrokedButtonComponent } from './stroked-button/stroked-button.component';
+export { StrokedButtonModule } from './stroked-button/stroked-button.module';
+import { UnderlineButtonComponent } from './underline-button/underline-button.component';
+export { UnderlineButtonModule } from './underline-button/underline-button.module';
 
-export { ButtonExamplesModule } from './button-examples.module';
-
-export { BUTTON_EXAMPLES } from './examples';
+export const BUTTON_EXAMPLES = [
+  BasicButtonComponent,
+  IconButtonComponent,
+  RaisedButtonComponent,
+  SizeableButtonComponent,
+  StrokedButtonComponent,
+  UnderlineButtonComponent,
+];
